feat:  createDocumentNotFoundError function

+++ src/controllers/JobExampleController.ts
@@ -9,6 +9,7 @@ import isValidUrl from "../utils/validUrlChecker.js";
 import { AuthenticatedRequest } from "../types/AuthenticatedRequest.js";
 import createValidationError from "../utils/createValidationError.js";
 import createCustomError from "../utils/createCustomError.js";
+import createDocumentNotFoundError from "../utils/createDocumentNotFoundError.js";
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
@@ -174,9 +175,8 @@ class JobExampleController extends BaseController {
       });
 
       if (!obtainedJobExample) {
-        throw createCustomError(
-          `JobExample with ID ${jobExampleId} not found`,
-          404
+        throw createDocumentNotFoundError(
+          `JobExample with ID ${jobExampleId} not found`
         );
       }
 
@@ -265,9 +265,8 @@ class JobExampleController extends BaseController {
       });
 
       if (!obtainedJobExample) {
-        throw createCustomError(
-          `JobExample with ID ${jobExampleId} not found`,
-          404
+        throw createDocumentNotFoundError(
+          `JobExample with ID ${jobExampleId} not found`
         );
       }
 
@@ -291,7 +290,7 @@ class JobExampleController extends BaseController {
               "../../uploads/image",
               picture
             );
-            if (fs.existsSync(filePath)) {
+            if (fs.existsSync(filePath) && fs.lstatSync(filePath).isFile()) {
               fs.unlinkSync(filePath);
             }
           });
@@ -303,7 +302,7 @@ class JobExampleController extends BaseController {
         if (files.videos) {
           (obtainedJobExample as any).videos.forEach((video: string) => {
             const filePath = path.join(__dirname, "../../uploads/video", video);
-            if (fs.existsSync(filePath)) {
+            if (fs.existsSync(filePath) && fs.lstatSync(filePath).isFile()) {
               fs.unlinkSync(filePath);
             }
           });
@@ -313,7 +312,7 @@ class JobExampleController extends BaseController {
         if (files.audios) {
           (obtainedJobExample as any).audios.forEach((audio: string) => {
             const filePath = path.join(__dirname, "../../uploads/audio", audio);
-            if (fs.existsSync(filePath)) {
+            if (fs.existsSync(filePath) && fs.lstatSync(filePath).isFile()) {
               fs.unlinkSync(filePath);
             }
           });
